feat(testing): expose a module dedicated to test env
We have now a second EntryPoint used by tsc|ngc which is ./testing/index.ts. In this angular module, we can find the whole list of components used in the standard part but without any logic, especially around OpenLayers. Thanks to that, we use this module `AngularOpenLayersTestingModule` in TestBed and prevent error due to instantiation of the Openlayers (which causes XMLHttpRequest for example...)
This modification leads to another big change around the build process. We need to expose different elements as a library (standard and testing module). In order to avoid some import as `import {AngularOpenLayersTestingModule} from 'ngx-openlayers/dist/testing'`, we choose to do some change in the build. To publish now, we have to use the file `publish.js` which will do the transformation required for us (package.json, publish...)

Closes https://github.com/quentin-ol/ngx-openlayers/issues/96